ENG-15003: Use netty SslContext instead of java SSLContext
Netty's SslContext can construct SSLEngine which is backed by OpenSsl.
OpenSsl can be considerably faster than using basic java TLS
implementation.

Reworked handshake logic to send server timestemp before handshake to
avoid the openssl engine from trying to decrypt the timestamp as an ssl
frame. This is because the unwrap method does not return that the
handshake is finished until the first call after the handshake is
finished. When the timestamp was received TLSHandshaker would try to
call unwrap on the data which would cause an exception and fail the
join.